requirements.txt
colorcet==3.0.1
csbdeep==0.7.3
dask==2023.4.1
imgaug==0.4.0
lib==4.0.0
magicgui==0.7.2
matplotlib==3.7.1
napari==0.4.17
numba==0.57.0
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
Pint==0.21
pip==23.0.1
prettytable==3.7.0
psygnal==0.9.0
pytest==7.3.1
PyYAML==6.0
PyYAML==6.0
QtPy==2.3.1
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==66.0.0
sympy==1.11.1
tifffile==2023.4.12
torch==2.0.0+cu118
torchsummary==1.5.1
torchvision==0.15.1+cu118
tqdm==4.65.0
requests
albumentations
torchio
torchsummary
colorcet
prettytable
tensorboardX
